+++ src/flinkSql/languageClient.ts
@@ -15,13 +15,14 @@ import { getSecretStorage } from "../storage/utils";
 import { getFlinkSQLLanguageServerOutputChannel } from "./logging";
 import { WebsocketTransport } from "./websocketTransport";
 
-const logger = new Logger("flinkSql.languageClient");
+const logger = new Logger("flinkSql.languageClient.Client");
 
-/** Initialize the FlinkSQL language client and connect to the language server websocket
+/** Initialize the FlinkSQL language client and connect to the language server websocket.
+ * Creates a WebSocket (ws), then on ws.onopen makes the WebsocketTransport class for server, and then creates the Client.
+ * Provides middleware for completions and diagnostics in ClientOptions
+ * @param url The URL of the language server websocket
+ * @param onWebSocketDisconnect Callback for WebSocket disconnection events
  * @returns A promise that resolves to the language client, or null if initialization failed
- * Prerequisites:
- * - User is authenticated with CCloud
- * - User has selected a compute pool
  */
 export async function initializeLanguageClient(
   url: string,
@@ -58,10 +59,11 @@ export async function initializeLanguageClient(
             { pattern: "**/*.flink.sql" },
           ],
           outputChannel: getFlinkSQLLanguageServerOutputChannel(),
+          synchronize: {
+            fileEvents: vscode.workspace.createFileSystemWatcher("**/*.flink.sql"),
+          },
+          progressOnInitialization: true,
           middleware: {
-            didOpen: (document, next) => {
-              return next(document);
-            },
             provideCompletionItem: async (document, position, context, token, next) => {
               const result: any = await next(document, position, context, token);
               if (result) {