+++ tests/e2e/specs/flinkStatement.spec.ts
@@ -1,5 +1,5 @@
 import { FrameLocator } from "@playwright/test";
-import { test } from "vscode-test-playwright";
+import { test } from "../baseTest";
 import { openConfluentExtension } from "./utils/confluent";
 import { login } from "./utils/confluentCloud";
 import {