+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this extension will be documented in this file.
 
 ## Unreleased
 
+### Changed
+
+- Updated ide-sidecar version to v0.216.0
+- Simplified subject and/or schema version hard deletion flow to include a warning dialog instead of
+  requiring the word `hard` to be typed.
+
 ## 1.4.0
 
 ### Added